#24 Refactor rust structure
