+++ src-tauri/src/types.rs
@@ -1,0 +1,2 @@
+pub mod enums;
+pub mod structs;
